[linker] Move XI/XM shared code from MobileSweepStep to a new CoreSweepStep class.
